Rade rezervacije

+++ src/main/java/com/edu/raf/NWP_Projekat/controllers/ReservationController.java
@@ -1,13 +1,24 @@
 package com.edu.raf.NWP_Projekat.controllers;
 
+import com.edu.raf.NWP_Projekat.model.Flight;
+import com.edu.raf.NWP_Projekat.model.Ticket;
 import com.edu.raf.NWP_Projekat.model.modelDTO.FlightDto;
 import com.edu.raf.NWP_Projekat.model.modelDTO.ReservationDto;
+import com.edu.raf.NWP_Projekat.model.modelDTO.ReservationsResponse;
+import com.edu.raf.NWP_Projekat.model.modelDTO.TicketDto;
+import com.edu.raf.NWP_Projekat.services.FlightService;
 import com.edu.raf.NWP_Projekat.services.ReservationService;
+import com.edu.raf.NWP_Projekat.services.TicketService;
+import com.edu.raf.NWP_Projekat.services.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.time.Duration;
+import java.time.LocalDate;
+import java.time.format.DateTimeFormatter;
 import java.util.List;
+import java.util.Locale;
 
 @RestController
 @RequestMapping("/api/reservations")
@@ -16,6 +27,12 @@ public class ReservationController {
     @Autowired
     private ReservationService reservationService;
 
+    @Autowired
+    private TicketService ticketService;
+
+    @Autowired
+    private UserService userService;
+
     @GetMapping("/all")
     public List<ReservationDto> getAllReservations(){
         return this.reservationService.getAllReservations();
@@ -23,6 +40,16 @@ public class ReservationController {
 
     @DeleteMapping(value = "/{id}")
     public ResponseEntity<?> deleteReservation(@PathVariable Long id){
+        ReservationDto reservationDto = this.reservationService.getById(id);
+        TicketDto ticket = this.ticketService.getById(reservationDto.getTicket());
+        DateTimeFormatter dateTimeFormatter = DateTimeFormatter.ofPattern("yyyy-MM-dd", Locale.UK);
+        LocalDate departureDate = LocalDate.parse(ticket.getDepartDate(), dateTimeFormatter);
+        LocalDate currentDate = LocalDate.now();
+        if(Duration.between(departureDate.atStartOfDay(), currentDate.atStartOfDay()).toDays() < 1){
+            //TODO Exception
+            System.out.println("-+--------------------------------------------------------------------");
+            return null;
+        }
         if(this.reservationService.deleteReservation(id)){
             return  ResponseEntity.ok().build();
         }
@@ -50,6 +77,11 @@ public class ReservationController {
             return ResponseEntity.ok(reservationDto);
         }
         return ResponseEntity.notFound().build();
+    }
+
+    @GetMapping(value = "/userReservations")
+    public List<ReservationsResponse> getById(@RequestParam(name = "username" ,required = true) String username){
+        return userService.getUserBookings(username);
     }
